New version with genericity
This version incorporates the new algorithmn, some benches and an example
to illustrate its use. The library seems not to behave linearithmicly at
really high (100M) elements. Flamegraphing does not seem to indicate a
bottlneck in the algorithm itself so _propably_ hardware-limited.

Some additional documentation would be needed to explain the library
more in-depth and will probably will never be written but let's have faith.